requirements.txt
absl-py==2.1.0
accelerate==0.32.1
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.5
aiosignal==1.3.1
aliyun-python-sdk-core==2.15.1
aliyun-python-sdk-kms==2.16.3
altair==5.3.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
attrs==23.2.0
audioread==3.0.1
auto_gptq==0.7.1
av==12.2.0
cachetools==5.3.3
certifi==2024.7.4
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cn2an==0.5.22
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.2.1
cramjam==2.8.3
crcmod==1.7
cryptography==42.0.8
ctranslate2==4.3.1
curl-cffi==0.5.7
cycler==0.12.1
Cython==0.29.37
datasets==2.20.0
decorator==5.1.1
dill==0.3.8
Distance==0.1.3
dnspython==2.6.1
editdistance==0.8.1
einops==0.8.0
email_validator==2.2.0
exllamav2==0.1.6
fastapi==0.111.0
fastapi-cli==0.0.4
faster-whisper==1.0.3
fastparquet==2024.5.0
ffmpeg-python==0.2.0
ffmpy==0.3.2
filelock==3.15.4
flatbuffers==24.3.25
fonttools==4.53.0
frozenlist==1.4.1
fsspec==2024.5.0
funasr==1.0.0
future==1.0.0
g2p-en==2.1.0
gast==0.6.0
gekko==1.2.1
google-api-core==2.19.1
google-api-python-client==2.136.0
google-auth==2.31.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.63.2
gradio==3.38.0
gradio_client==0.8.1
grpcio==1.64.1
h11==0.14.0
hdbscan==0.8.37
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.23.4
humanfriendly==10.0
hydra-core==1.3.2
idna==3.7
importlib_metadata==8.0.0
inflect==7.3.1
intel-openmp==2021.4.0
iso8601==2.1.0
jaconv==0.3.4
jamo==0.4.1
jieba==0.42.1
jieba_fast==0.53
Jinja2==3.1.4
jmespath==0.10.0
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
kaldiio==2.18.0
kick-chat==1.0.0
kiwisolver==1.4.5
LangSegment==0.3.3
librosa==0.9.2
lightning-utilities==0.11.3.post0
linkify-it-py==2.0.3
llvmlite==0.43.0
Markdown==3.6
markdown-it-py==2.2.0
MarkupSafe==2.1.5
matplotlib==3.9.1
mdit-py-plugins==0.3.3
mdurl==0.1.2
mkl==2021.4.0
modelscope==1.10.0
more-itertools==10.3.0
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
networkx==3.3
ninja==1.11.1.1
nltk==3.8.1
numba==0.60.0
numpy==1.26.4
oauth2client==4.1.3
omegaconf==2.3.0
onnxruntime==1.18.1
openai-whisper==20231117
optimum==1.21.1
orjson==3.10.6
oss2==2.18.6
packaging==24.1
pandas==2.2.2
peft==0.11.1
pillow==10.4.0
platformdirs==4.2.2
pooch==1.8.2
proces==0.1.7
proto-plus==1.24.0
protobuf==4.25.3
psutil==6.0.0
py3langid==0.2.2
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pycparser==2.21
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pynput==1.7.7
pyopenjtalk==0.3.3
pyparsing==3.1.2
pypinyin==0.51.0
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytorch-lightning==2.3.2
pytorch-wpe==0.0.1
pytest==8.3.3
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
rel==0.4.9
requests==2.32.3
resampy==0.4.3
rich==13.7.1
rouge==1.0.1
rpds-py==0.18.1
rsa==4.9
safetensors==0.4.3
scikit-learn==1.5.1
scipy==1.14.0
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
simplejson==3.19.2
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
SpeechRecognition==3.10.4
starlette==0.37.2
sympy==1.12.1
tbb==2021.13.0
tensorboard==2.17.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.19.1
tomli==2.0.1
toolz==0.12.1
tqdm==4.66.4
transformers==4.42.3
twitchio==2.10.0
typeguard==4.3.0
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
uc-micro-py==1.0.3
ujson==5.10.0
umap==0.1.1
uritemplate==4.1.1
urllib3==2.2.2
uvicorn==0.30.1
watchfiles==0.22.0
websocket-client==1.6.1
websockets==11.0.3
Werkzeug==3.0.3
wordsegment==1.3.1
xxhash==3.4.1
yapf==0.40.2
yarl==1.9.4
zipp==3.19.2
requests_oauthlib==2.0.0
robotframework==7.1.1
robotframework-pythonlibcore==4.4.1
robotframework-seleniumlibrary==6.6.1